+++ tests/test_gates.py
@@ -117,7 +117,7 @@ def test_gate_unitary_correct(op, op_name):
     unitary = np.zeros((2 ** wires, 2 ** wires), dtype=np.complex128)
 
     for i, input in enumerate(itertools.product([0, 1], repeat=wires)):
-        out = output(input)
+        out = output(np.array(input))
         unitary[:, i] = out
 
     unitary_expected = op(*p, wires=range(wires)).matrix
@@ -162,7 +162,7 @@ def test_inverse_unitary_correct(op, op_name):
     unitary = np.zeros((2 ** wires, 2 ** wires), dtype=np.complex128)
 
     for i, input in enumerate(itertools.product([0, 1], repeat=wires)):
-        out = output(input)
+        out = output(np.array(input))
         unitary[:, i] = out
 
     unitary_expected = op(*p, wires=range(wires)).inv().matrix
@@ -215,7 +215,7 @@ def test_arbitrary_unitary_correct():
     unitary = np.zeros((2 ** wires, 2 ** wires), dtype=np.complex128)
 
     for i, input in enumerate(itertools.product([0, 1], repeat=wires)):
-        out = output(input)
+        out = output(np.array(input))
         unitary[:, i] = out
 
     assert np.allclose(unitary, random_unitary)
@@ -236,7 +236,7 @@ def test_arbitrary_inv_unitary_correct():
     unitary = np.zeros((2 ** wires, 2 ** wires), dtype=np.complex128)
 
     for i, input in enumerate(itertools.product([0, 1], repeat=wires)):
-        out = output(input)
+        out = output(np.array(input))
         unitary[:, i] = out
 
     random_unitary_inv = random_unitary.conj().T